Minor update on simulate.py documentation
Pat and I decided that it might be a good idea to make more explicitly in the documentation that the package can calculate biases using the Fisher Formalism. I have included this in the `programs.rst` part of the documentation where `simulate.py` is described, I basically describe the new command line argument `--calculate-bias`. What do you think of this change @dkirkby? Is there any other part of the documentation where we could mention the capability of calculating biases?